infrastructure/common-images/workspaces-base/backend/requirements.txt
uritemplate==4.1.1
psycopg2-binary
django-dry-rest-permissions==1.2.0
Pillow==9.0.1
bg-atlasapi==1.0.2

fastapi==0.77.1
uvicorn==0.17.6
